refactor: make nonDeterministicLimitedRandReadSeeker re-usable across packages
